Bugnot

+++ Libraries/PCA9685/PCA9685.cpp
@@ -162,8 +162,8 @@ bool PCA9685::setPWM(uint8_t channel, uint16_t on, uint16_t off) {
     std::cout << "setPWM ch=" << (int)channel
             << " on=" << on
             << " off=" << off
-            << " reg=" << reg 
-            << " LED0_ON_L=" << LED0_ON_L
+            << " reg=" << (int)reg 
+            << " LED0_ON_L=" << (int)LED0_ON_L
             << std::endl;
 
     return writeBlock(reg, data, sizeof(data));
